🎨 #36 style: todoitem 스타일링
